Move components to its own folder in ui package
