+++ app/src/main/java/com/rebalance/ui/component/ExpandableList.kt
@@ -1,4 +1,4 @@
-package com.rebalance.ui.components
+package com.rebalance.ui.component
 
 import android.content.Context
 import android.graphics.BitmapFactory
